// src/presentation/container/components/custom/TabViewItemTwoScreen/type.ts
import {ViewComponent, ViewStyle, TextStyle} from 'react-native';

export type CustomTabViewItemProps = {
  screen1?: any;
  screen2?: any;
  screen3?: any;
  title1?: string;
  title2?: string;
  title3?: string;
  viewStyle?: ViewStyle;
  tabStyle?: ViewStyle;
  titleStyle?: TextStyle;
  titleStyle2?: TextStyle;

};